# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hasanaa e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subcategories:
    get:
      operationId: listSubcategories
      tags: [subcategories]
      summary: List subcategories, optionally filtered by categoryId
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of subcategories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subcategory"
    post:
      operationId: createSubcategory
      tags: [subcategories]
      summary: Create a subcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubcategoryBody"
      responses:
        "201":
          description: Subcategory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
  /subcategories/{id}:
    get:
      operationId: getSubcategory
      tags: [subcategories]
      summary: Get a subcategory with its products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subcategory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubcategoryWithProducts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSubcategory
      tags: [subcategories]
      summary: Update a subcategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubcategoryBody"
      responses:
        "200":
          description: Subcategory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    delete:
      operationId: deleteSubcategory
      tags: [subcategories]
      summary: Delete a subcategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subcategory deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Subcategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - categoryId
        - imageUrl
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    SubcategoryWithProducts:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required:
        - id
        - name
        - slug
        - categoryId
        - imageUrl
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
        - products
    CreateSubcategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        categoryId:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - slug
        - categoryId
    UpdateSubcategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        slug:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        sortOrder:
          type: ["integer", "null"]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        originalPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        subcategoryId:
          type: ["integer", "null"]
        subcategoryName:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - originalPrice
        - imageUrl
        - images
        - categoryId
        - categoryName
        - subcategoryId
        - subcategoryName
        - inStock
        - isFeatured
        - sizes
        - colors
        - createdAt
        - updatedAt
...